+++ release_notes/release_1.2/release_notes.md
@@ -5,6 +5,6 @@
 This release includes:
 
 - examples of invalid XML files to test XSD files against ✓
-- unit tests for Schemata code
+- unit tests for Schemata code ✓
 - a set of N semantic tags
 - examples, a specification, and an XSD file for Reader XML